refactor: path updates and argument changes as per: #00dfc42b.
