+++ dialogy/plugins/__init__.py
@@ -1,11 +1,6 @@
-from dialogy.plugins.postprocess.text.slot_filler.rule_slot_filler import (
-    RuleBasedSlotFillerPlugin,
-)
-from dialogy.plugins.postprocess.text.voting.intent_voting import VotePlugin
-from dialogy.plugins.preprocess.text.calibration import (
-    WERCalibrationConfig,
-    WERCalibrationPlugin,
-)
-from dialogy.plugins.preprocess.text.duckling_plugin import DucklingPlugin
-from dialogy.plugins.preprocess.text.list_entity_plugin import ListEntityPlugin
-from dialogy.plugins.preprocess.text.merge_asr_output import MergeASROutputPlugin
+from dialogy.plugins.text.calibration import WERCalibrationConfig, WERCalibrationPlugin
+from dialogy.plugins.text.duckling_plugin import DucklingPlugin
+from dialogy.plugins.text.list_entity_plugin import ListEntityPlugin
+from dialogy.plugins.text.merge_asr_output import MergeASROutputPlugin
+from dialogy.plugins.text.slot_filler.rule_slot_filler import RuleBasedSlotFillerPlugin
+from dialogy.plugins.text.voting.intent_voting import VotePlugin
