+++ dialogy/plugins/text/list_entity_plugin.py
