+++ dialogy/plugins/text/normalize_utterance.py
@@ -46,7 +46,7 @@ def is_each_element(
 
     .. ipython:: python
 
-        from dialogy.plugins.preprocess.text.normalize_utterance import is_each_element
+        from dialogy.plugins.text.normalize_utterance import is_each_element
 
         is_each_element(str, ["this", "returns", "False", "cuz:", False])
         is_each_element(str, ["this", "returns", "True", "cuz:", "all", "str"])
@@ -81,7 +81,7 @@ def is_utterance(maybe_utterance: Any, key: str = const.TRANSCRIPT) -> bool:
 
     .. ipython:: python
 
-        from dialogy.plugins.preprocess.text.normalize_utterance import is_utterance
+        from dialogy.plugins.text.normalize_utterance import is_utterance
 
         # 1. :code:`List[List[Dict[str, str]]]`
         is_utterance([[{"transcript": "this"}, {"transcript": "works"}]])
@@ -119,7 +119,7 @@ def is_unsqueezed_utterance(maybe_utterance: Any, key: str = const.TRANSCRIPT) -
 
     .. ipython:: python
 
-        from dialogy.plugins.preprocess.text.normalize_utterance import is_unsqueezed_utterance
+        from dialogy.plugins.text.normalize_utterance import is_unsqueezed_utterance
 
         # 1. This fails
         is_unsqueezed_utterance([[{"transcript": "this"}, {"transcript": "works"}]])
@@ -149,7 +149,7 @@ def is_list_of_string(maybe_utterance: Any) -> bool:
 
     .. ipython:: python
 
-        from dialogy.plugins.preprocess.text.normalize_utterance import is_list_of_string
+        from dialogy.plugins.text.normalize_utterance import is_list_of_string
         is_list_of_string(["this", "works"])
 
     :param maybe_utterance: Arbitrary input.
@@ -196,7 +196,7 @@ def normalize(maybe_utterance: Any, key: str = const.TRANSCRIPT) -> List[str]:
 
     .. ipython:: python
 
-        from dialogy.plugins.preprocess.text.normalize_utterance import normalize
+        from dialogy.plugins.text.normalize_utterance import normalize
 
         normalize([[{"transcript": "hi"}]])
         normalize([[{"transcript": "hello"}], [{"transcript": "world"}]])
