+++ dialogy/plugins/text/voting/intent_voting.py
