+++ tests/plugin/text/slot_filler/test_rule_slot_filler.py
@@ -38,7 +38,7 @@ def test_slot_filling() -> None:
     slot_filler = RuleBasedSlotFillerPlugin(rules=rules, access=access)()
 
     # Create a mock `workflow`
-    workflow = Workflow(preprocessors=[], postprocessors=[slot_filler])
+    workflow = Workflow([slot_filler])
 
     # ... a mock `Intent`
     intent = Intent(name=intent_name, score=0.8)
@@ -77,7 +77,7 @@ def test_slot_no_fill() -> None:
     slot_filler = RuleBasedSlotFillerPlugin(rules=rules, access=access)()
 
     # Create a mock `workflow`
-    workflow = Workflow(preprocessors=[], postprocessors=[slot_filler])
+    workflow = Workflow([slot_filler])
 
     # ... a mock `Intent`
     intent = Intent(name=intent_name, score=0.8)
@@ -116,7 +116,7 @@ def test_slot_dual_fill() -> None:
     slot_filler = RuleBasedSlotFillerPlugin(rules=rules, access=access)()
 
     # Create a mock `workflow`
-    workflow = Workflow(preprocessors=[], postprocessors=[slot_filler])
+    workflow = Workflow([slot_filler])
 
     # ... a mock `Intent`
     intent = Intent(name=intent_name, score=0.8)
@@ -166,7 +166,7 @@ def test_slot_filling_multiple() -> None:
     )()
 
     # Create a mock `workflow`
-    workflow = Workflow(preprocessors=[], postprocessors=[slot_filler])
+    workflow = Workflow([slot_filler])
 
     # ... a mock `Intent`
     intent = Intent(name=intent_name, score=0.8)
@@ -213,7 +213,7 @@ def test_slot_competition() -> None:
     slot_filler = RuleBasedSlotFillerPlugin(rules=rules, access=access)()
 
     # Create a mock `workflow`
-    workflow = Workflow(preprocessors=[], postprocessors=[slot_filler])
+    workflow = Workflow([slot_filler])
 
     # ... a mock `Intent`
     intent = Intent(name=intent_name, score=0.8)
@@ -254,7 +254,7 @@ def test_incorrect_access_fn() -> None:
     access = 5
 
     slot_filler = RuleBasedSlotFillerPlugin(rules=rules, access=access)()
-    workflow = Workflow(preprocessors=[], postprocessors=[slot_filler])
+    workflow = Workflow([slot_filler])
     intent = Intent(name="intent", score=0.8)
 
     body = "12th december"
@@ -277,7 +277,7 @@ def test_missing_access_fn() -> None:
     This test shows that the plugin needs an `access` provided or else it raises a type error.
     """
     slot_filler = RuleBasedSlotFillerPlugin(rules=rules)()
-    workflow = Workflow(preprocessors=[], postprocessors=[slot_filler])
+    workflow = Workflow([slot_filler])
     intent = Intent(name="intent", score=0.8)
 
     body = "12th december"
