+++ tests/workflow/test_workflow.py
@@ -7,7 +7,7 @@ def test_workflow_get_input() -> None:
     """
     Basic initialization.
     """
-    workflow = Workflow(preprocessors=[], postprocessors=[])
+    workflow = Workflow([])
     assert workflow.input == {}, "workflow.get_input() is a dict()."
 
 
@@ -15,18 +15,9 @@ def test_workflow_set_output() -> None:
     """
     We can set output and input as anything.
     """
-    workflow = Workflow(preprocessors=[], postprocessors=[])
+    workflow = Workflow([])
     workflow.output = 10
     assert workflow.output == 10, "workflow.get_output() should == 10."
-
-
-def test_workflow_load_model_error() -> None:
-    """
-    `load_model` has to be defined by the subclass.
-    """
-    workflow = Workflow(preprocessors=[], postprocessors=[])
-    with pytest.raises(NotImplementedError):
-        workflow.load_model()
 
 
 def test_workflow_postprocessors_not_list_error() -> None:
@@ -34,15 +25,7 @@ def test_workflow_postprocessors_not_list_error() -> None:
     postprocessrors should be of type [`PluginFn`](../../plugin/test_plugins.html).
     """
     with pytest.raises(TypeError):
-        _ = Workflow(preprocessors=[], postprocessors=10)  # type: ignore
-
-
-def test_workflow_preprocessors_not_list_error() -> None:
-    """
-    preprocessors should be of type [`PluginFn`](../../plugin/test_plugins.html)
-    """
-    with pytest.raises(TypeError):
-        _ = Workflow(preprocessors=10, postprocessors=[])  # type: ignore
+        _ = Workflow(10)  # type: ignore
 
 
 def test_workflow_run() -> None:
@@ -56,7 +39,7 @@ def test_workflow_run() -> None:
     def mock_preproc(w) -> None:
         w.input = 20
 
-    workflow = Workflow(preprocessors=[mock_preproc], postprocessors=[mock_postproc])
+    workflow = Workflow([mock_preproc, mock_postproc])
 
     # initially the value of workflow.input is = 2
     workflow.run(2)
@@ -67,28 +50,12 @@ def test_workflow_run() -> None:
     assert workflow.output == 10, "workflow.get_output() should be 10."
 
 
-def test_child_class_inference_not_implemented_error() -> None:
-    """
-    Sub-classes should implement the `inference` method.
-    """
-
-    class Wkflw(Workflow):
-        pass
-
-    w = Wkflw(preprocessors=[], postprocessors=[])
-    with pytest.raises(NotImplementedError):
-        w.inference()
-
-
 def test_not_callable_pre_post_processors_type_error() -> None:
     """"""
-    workflow = Workflow(preprocessors=[2], postprocessors=[3])  # type: ignore
+    workflow = Workflow([2, 3])  # type: ignore
 
     with pytest.raises(TypeError):
-        workflow.preprocess()
-
-    with pytest.raises(TypeError):
-        workflow.postprocess()
+        workflow.execute()
 
 
 def test_workflow_run_debug_mode() -> None:
@@ -102,9 +69,7 @@ def test_workflow_run_debug_mode() -> None:
     def mock_preproc(w):
         w.input = 20
 
-    workflow = Workflow(
-        preprocessors=[mock_preproc], postprocessors=[mock_postproc], debug=True
-    )
+    workflow = Workflow([mock_preproc, mock_postproc], debug=True)
 
     workflow.run(2)
 
@@ -123,7 +88,7 @@ def test_workflow_flush() -> None:
     def mock_preproc(w) -> None:
         w.input = 20
 
-    workflow = Workflow(preprocessors=[mock_preproc], postprocessors=[mock_postproc])
+    workflow = Workflow([mock_preproc, mock_postproc])
 
     # initially the value of workflow.input is = 2
     workflow.run(2)
@@ -148,6 +113,6 @@ def test_workflow_json() -> None:
     def mock_preproc(w) -> None:
         w.input = 20
 
-    workflow = Workflow(preprocessors=[mock_preproc], postprocessors=[mock_postproc])
+    workflow = Workflow([mock_preproc, mock_postproc])
     workflow.run(1)
     assert workflow.json() == {"input": 20, "output": 10}
